move errors.go into interceptor package
